Update angular-directive-percentage.js
I removed the check valid check in formatViewDataIntoModelData as this is the new value and we are about to check it's validity and it was interfering with the required directive. Also had a issue with multiple instances of the directive on the same page. I tried an isolate scope but it was still pulling in min and max values from the wrong element. I moved all of the logic into the link method so that each element could store it's own min, max variable values. It seems to be performing as desired now.